+++ docusaurus.config.ts
@@ -157,14 +157,14 @@ const config: Config = {
         },
 
         // Announcement Bar
-        announcementBar: {
-            id: 'internal-prototype',
-            content:
-                '<b>Thank you for visiting the SAP Architecture Center. Your <a href="https://github.com/SAP/architecture-center/discussions" target="_blank">feedback</a> is important to us!</b>',
-            backgroundColor: '#0053CB',
-            textColor: '#FFFFFF',
-            isCloseable: true,
-        },
+        // announcementBar: {
+        //     id: 'internal-prototype',
+        //     content:
+        //         '<b>Thank you for visiting the SAP Architecture Center. Your <a href="https://github.com/SAP/architecture-center/discussions" target="_blank">feedback</a> is important to us!</b>',
+        //     backgroundColor: '#0053CB',
+        //     textColor: '#FFFFFF',
+        //     isCloseable: true,
+        // },
         navbar: {
             title: 'Architecture Center',
             hideOnScroll: false,
