--- src/sections/TrustedTecPartnersSection.module.css
@@ -1,194 +1,0 @@
-.trustedTecPartnersSection {
-    background-color: #c8c8c8;
-    height: auto;
-}
-
-.innerWrapper {
-    width: 1440px;
-    height: 480px;
-    margin: 0 auto;
-    background-image: url("/img/landingPage/worldmap_1440x424_light.webp");
-    background-repeat: no-repeat;
-    background-size: cover;
-    background-position: center;
-    padding: 30px 24px 10px;
-}
-
-:root[data-theme='dark'] .trustedTecPartnersSection {
-    background-color: #000;
-}
-
-:root[data-theme='dark'] .innerWrapper {
-    background-image: url("/img/landingPage/worldmap_1440x424_dark.webp")
-}
-
-.logoWrapper {
-    flex: 0 0 auto;
-    width: 220px;
-    height: 100px;
-    display: flex;
-    align-items: center;
-    justify-content: center;
-    padding: 0 24px;
-    box-sizing: border-box;
-    flex-shrink: 0;
-    margin: 0 20px;
-    transition: transform 0.4s cubic-bezier(0.175, 0.885, 0.32, 1.275);
-}
-
-.logoWrapper:hover {
-    transform: scale(1.1) translateY(-5px);
-}
-
-.logoImg {
-    height: 80%;
-    width: 80%;
-    max-height: 80px;
-    max-width: 200px;
-    display: block;
-    object-fit: contain;
-    transition: transform 0.3s ease;
-}
-
-.logoImg:hover {
-    transform: scale(1.2);
-}
-
-.title {
-    font-size: 2.5rem;
-    font-weight: 700;
-    color: var(--ifm-font-color-base);
-    margin-bottom: 1rem;
-}
-
-.subtitle {
-    font-size: 1.25rem;
-    color: var(--ifm-color-content-secondary);
-    max-width: 700px;
-    margin: 0;
-    line-height: 1.6;
-}
-
-.carouselLogo {
-    padding-top: clamp(40px, 5vw, 80px);
-    position: relative;
-    overflow: hidden;
-}
-
-/* Smooth infinite scroll container */
-.carouselTrack {
-    display: flex;
-    width: max-content;
-    animation: scroll 60s linear infinite;
-    will-change: transform;
-}
-
-@keyframes scroll {
-    0% {
-        transform: translateX(0);
-    }
-
-    100% {
-        transform: translateX(-33.33333%);
-        /* Move exactly by 1/3 since we have 3 identical sets */
-    }
-}
-
-/* Carousel fade elements removed to eliminate "frame" look */
-
-.logoList {
-    display: none;
-}
-
-:root[data-theme='dark'] .title,
-:root[data-theme='dark'] .subtitle {
-    color: var(--text-color-dark-mode);
-}
-
-/* Tablet */
-@media (max-width: 1024px) {
-    .logoImg {
-        max-width: 140px;
-        max-height: 70px;
-    }
-
-    .carouselLogo {
-        padding: 0 20px;
-    }
-
-    .logoWrapper {
-        padding: 0 20px;
-    }
-
-    .innerWrapper {
-        width: auto;
-        height: auto;
-    }
-}
-
-/* Mobile */
-@media (max-width: 600px) {
-    .title {
-        font-size: var(--sapFontHeader3Size);
-    }
-
-    .subtitle {
-        font-size: var(--sapFontHeader4Size);
-    }
-
-    .logoImg {
-        max-width: 100px;
-        max-height: 60px;
-    }
-
-    .carouselLogo {
-        display: none;
-    }
-
-    .logoList {
-        display: grid;
-        grid-template-columns: repeat(2, auto);
-        gap: 20px 40px;
-        width: max-content;
-        margin: 30px auto 20px;
-    }
-
-    .innerWrapper {
-        width: auto;
-        height: auto;
-    }
-
-}
-
-/* Small mobile */
-@media (max-width: 430px) {
-    .title {
-        font-size: var(--sapFontHeader4Size);
-    }
-
-    .subtitle {
-        font-size: var(--sapFontHeader5Size);
-    }
-
-    .logoImg {
-        max-width: 80px;
-        max-height: 50px;
-    }
-
-    .carouselLogo {
-        padding: 0 20px;
-    }
-
-    .logoList {
-        display: grid;
-        grid-template-columns: repeat(2, auto);
-        gap: 20px 40px;
-        width: max-content;
-        margin: 30px auto 20px;
-    }
-
-    .innerWrapper {
-        width: auto;
-        height: auto;
-    }
-}
